Nested components. Part 1
