+++ src/evolution_front/src/actions/room.actions.js
@@ -15,6 +15,7 @@ export const roomActions = {
   list,
   get,
   user,
+  getUsers,
 };
 
 function create(roomName) {
@@ -336,5 +337,30 @@ function user() {
   }
   function failure(error) {
     return { type: roomConstants.USER_FAILURE, error };
+  }
+}
+
+function getUsers(roomUid) {
+  return async (dispatch) => {
+    dispatch(request(roomUid));
+
+    return roomService.getUsers(roomUid).then(
+      (users) => {
+        return dispatch(success(users));
+      },
+      (error) => {
+        return Promise.reject(dispatch(failure(error)));
+      }
+    );
+  };
+
+  function request(roomUid) {
+    return { type: roomConstants.GET_USERS_REQUEST, roomUid };
+  }
+  function success(users) {
+    return { type: roomConstants.GET_USERS_SUCCESS, users };
+  }
+  function failure(error) {
+    return { type: roomConstants.GET_USERS_FAILURE, error };
   }
 }
