+++ src/evolution_front/src/constants/room.constants.js
@@ -51,4 +51,8 @@ export const roomConstants = {
   USER_REQUEST: 'ROOM_USER_REQUEST',
   USER_SUCCESS: 'ROOM_USER_SUCCESS',
   USER_FAILURE: 'ROOM_USER_FAILURE',
+
+  GET_USERS_REQUEST: 'ROOM_GET_USERS_REQUEST',
+  GET_USERS_SUCCESS: 'ROOM_GET_USERS_SUCCESS',
+  GET_USERS_FAILURE: 'ROOM_GET_USERS_FAILURE',
 };
